pyproject.toml
[project]
name = "openai-python-translator"
version = "0.1.0"
description = "A simple OpenAI-based translator tool"
authors = [
    { name = "Dagne", email = "dagne.venskus@gmail.com" }
]
dependencies = [
    "annotated-types==0.7.0",
    "anyio==4.9.0",
    "certifi==2025.4.26",
    "colorama==0.4.6",
    "distro==1.9.0",
    "h11==0.16.0",
    "httpcore==1.0.9",
    "httpx==0.28.1",
    "idna==3.10",
    "jiter==0.10.0",
    "openai==1.79.0",
    "pydantic==2.11.4",
    "pydantic_core==2.33.2",
    "python-dotenv==1.1.0",
    "sniffio==1.3.1",
    "tqdm==4.67.1",
    "typing-inspection==0.4.0",
    "typing_extensions==4.13.2"
]

[tool.setuptools]
packages = ["translator"]

[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"
